Second Excel Project
This project is all about coffee sales in three different countries which are The United Kingdom, The United States and Ireland.
The dataset for this project was in three different worksheets within an excel file. The dataset has three tables which are the orders table, customers table and
the products table. The orders table had columns such as Order ID, Customer ID, Product ID and so on. Some of the columns had data already while some of the columns
were empty with their data present in other tables within the worksheet. The first thing I did in this project was data gathering.
I used the xlookup function in excel to populate the customers data in the orders table from the customers table and I carried out the same process
on the products data. After gathering the necessary data, I populated the sales column using the formula (unit price*quantity). The names of the coffee type and
roast type in the dataset were abbreviated, so I created new columns and I used the IF statement in excel to generate the full names of the coffee type and roast type.
After this step in the project, I changed the month (figure) in the order date column to month (abbreviated alphabets) to avoid confusions between the American and
European date formats. After data gathering and data formatting, I did some analysis on the dataset with the use of pivot tables and pivot charts. When I finished with the analysis, I created a coffee sales dashboard
with all the charts that I created during the analysis, I also added some slicers to my dashboard and a timeline to filter through the order date.