Third Excel Project
This project is about bike sales. The dataset of this project was in a github repository, so I downloaded the ZIP file for that repository
and I had access to the dataset. This dataset had columns such as Unique id, Marital Status, Gender ,Income, Commute Distance, Purchased Bike and so on.
The first I did in this project was to create a working sheet which is a copy of the original data while leaving the original/raw dataset in another sheet.
I did this so that I would still have access to the original data in case of any mistakes during the analytics process. The next thing I did to this dataset was to
remove duplicates. After this step in the project, the next thing I did was data formatting. Some of the things I did was to change the marital status symbols and
gender symbols to their full meanings. The next thing I did was to add a new column to the dataset called "Age bracket". I created the new column to group the
different ages in the "Age" column into a specific age bracket. After this stage in the project, the next thing I did was to do some analysis with the use of
pivot tables and pivot charts and then I created a dashboard with the pivot charts that I generated during my analysis. Lastly I added some slicers to the dashboard
in order to visualize the results for different situations.